피보나치 수 2

+++ src/Backjoon2747.java
@@ -6,10 +6,10 @@ public class Backjoon2747 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int input = sc.nextInt();
-        ArrayList<Integer> list = new ArrayList<Integer>();
+        ArrayList<Long> list = new ArrayList<Long>();
         
-        list.add(0);
-        list.add(1);
+        list.add((long)0);
+        list.add((long)1);
         
         for(int i = 2; i <= input; i++) {
             list.add(i, list.get(i-1) + list.get(i-2));
